Created Lazy implementation of UserReplies.
Created LazyUserReplies class that extends basic implementation of UserReplies interface.
Created a currentReply counter to determine the current position in the list of replies.
Added repliesList for storing objects of the Replies class.
Added replyList for storing the current list of replies.
Created constructor LazyUserReplies(Replies replies) to initialize the variables currentReply, repliesList, replyList.
Implemented getNextReplies method to return the following replies from replyList.
Made a temporary implementation of the getSentReplies() method.
Implemented addReplies(Replies replies) method to add Replies type objects to repliesList.
Implemented the getCurrentReply() method to get the index of the current position in the replies list.