The base implementation of RepliesManager has been reworked.
The BaseRepliesManager class is redesigned so that it no longer monitors and changes the order of the replies in the list, stores UserReplies for a specific user and returns suitable replies.
Removed nextReply and sentReplies dictionaries for storing next and sent replies.
Removed getNextReply() and getSentReplies() methods.
Removed abstract methods setNextReplies(), setSentReplies(), findRepliesByData(BaseIncMessage incMessage) and findRepliesByReply(BaseIncMessage incMessage, Reply reply reply), as their implementation would be unnecessary.
Added users dictionary to store UserReplies for a specific user.
Added abstract method setUserReplies(Replies replies): UserReplies.
Redesigned the getSuitableReplies(BaseIncMessage incMessage) method so that it no longer keeps track of the order of replies for a specific user and does not change it.